Authentification via header

+++ services/FileService/app.py
@@ -2,10 +2,12 @@ import json
 import os
 from datetime import datetime, date
 
+from flask_jwt_extended import jwt_required, get_jwt_identity, JWTManager
+from werkzeug.utils import secure_filename
 from depot.io.interfaces import FileStorage
 from depot.manager import DepotManager
 from depot.fields.sqlalchemy import UploadedFileField
-from flask import Flask, request, Response, jsonify
+from flask import Flask, request, Response
 from flask_sqlalchemy import SQLAlchemy
 from flasgger import Swagger
 
@@ -13,6 +15,13 @@ from flasgger import Swagger
 app = Flask(__name__)
 app.config['DEBUG'] = True
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:////tmp/test3.db'
+app.config['JWT_TOKEN_LOCATION'] = 'headers'
+app.config['JWT_HEADER_NAME'] = 'Authorization'
+app.config['JWT_HEADER_TYPE'] = 'Bearer'
+
+# security
+app.config['JWT_SECRET_KEY'] = 'secret'  # Change this!
+jwt = JWTManager(app)
 
 db = SQLAlchemy(app)
 """Dokumentation & Testanfragen zu finden unter /apidocs/"""
@@ -43,6 +52,9 @@ class FileEncoder(json.JSONEncoder):
 
 
 class File1(db.Model):
+    # todo list user freigeschaltet, für alle id (-1)
+    # todo put für dateiname and freigaben
+    # todo get für freigegebene Dateien mit query string parameter von get eigene zu unterscheiden
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(128), index=True, unique=True)
     size = db.Column(db.Integer)
@@ -68,18 +80,14 @@ def hello_world():
 
 
 @app.route("/files", methods=["GET"])
+@jwt_required
 def get_files():
     """
     Get all files.
     ---
     parameters:
-      - name: username
-        in: query
-        description: authentication token of user
-        required: true
-        type: string
       - name: usertoken
-        in: query
+        in: header
         description: authentication token of user
         required: true
         type: string
@@ -103,22 +111,34 @@ def get_files():
 
     """
     # get username
-    username = request.args.get("username")
-    usertoken = request.args.get("usertoken")
+    username = get_jwt_identity()
     if not username:
-        return not_authenticated(msg="Parameter 'username' is missing.")
-    elif not usertoken:
-        return not_authenticated(msg="Parameter 'usertoken' is missing.")
-    # authenticate with token
-    elif not authenticate(usertoken):
-        return not_authenticated(msg="No valid token.")
-    # get files from db
+        return not_authenticated()
+    # else: get files from db
     db_response = File1.query.filter_by(user=username).all()
     js = json.dumps(db_response, cls=FileEncoder)
     return Response(js, status=200, mimetype='application/json')
 
 
+
+
+def generate_secure_and_unique_filename(file):
+    secure_name = secure_filename(file.filename)
+    if not secure_name:
+        secure_name = "file_without_name"
+    if not File1.query.filter_by(name=secure_name).scalar():
+        return secure_name
+    i = 1
+    while File1.query.filter_by(name=secure_name).scalar():
+        secure_name = secure_name.split(".")[0] + str(i) + "." + secure_name.split(".")[1]
+        i += 1
+        if i > 1000:
+            raise Exception()
+    return secure_name
+
+
 @app.route("/files", methods=["POST"])
+@jwt_required
 def add_file():
     """
     Add a new file
@@ -129,8 +149,8 @@ def add_file():
     consumes:
       - multipart/form-data
     parameters:
-      - name: username
-        in: formData
+      - name: usertoken
+        in: header
         description: user name
         required: true
         type: string
@@ -156,22 +176,26 @@ def add_file():
               type: string
      """
     file = request.files["file_content"]
-    username = request.form.get("username")
+    username = get_jwt_identity()
+    if not username:
+        return not_authenticated()
     if file:
         # depot = DepotManager.get()
         # fileid = depot.create(file)
         # stored_file = depot.get(fileid)
         file.seek(0, os.SEEK_END)
         file_length = file.tell()
-        file_object = File1(file.filename, file_length, username, file)
+        secure_filename = generate_secure_and_unique_filename(file)
+        file_object = File1(secure_filename, file_length, username, file)
         db.session.add(file_object)
         db.session.commit()
         js = json.dumps(file_object, cls=FileEncoder)
         return Response(js, status=201, mimetype='application/json')
 
 
-@app.route("/files/<id>", methods=["DELETE"])
-def delete_file(id):
+@app.route("/files/<file_id>", methods=["DELETE"])
+@jwt_required
+def delete_file(file_id):
     """
     Delete a File by its id
     ---
@@ -180,14 +204,24 @@ def delete_file(id):
         in: path
         type: integer
         required: True
+      - name: usertoken
+        in: header
+        description: authentication token of user
+        required: true
+        type: string
     responses:
       200:
         description: Successfully deleted a file from db
       404:
         description: Not Found
     """
-    file = File1.query.get(id)
+    username = get_jwt_identity()
+    if not username:
+        return not_authenticated()
+    file = File1.query.get(file_id)
     if file:
+        if file.user != username:
+            return not_authorized()
         depot = DepotManager.get()
         depot.delete(file.fileid)
         db.session.delete(file)
@@ -200,8 +234,9 @@ def delete_file(id):
         return not_found()
 
 
-@app.route("/files/<id>", methods=["GET"])
-def get_single_file(id):
+@app.route("/files/<file_id>", methods=["GET"])
+@jwt_required
+def get_single_file(file_id):
     """
     Get one specific file by its id.
     ---
@@ -211,6 +246,11 @@ def get_single_file(id):
         description: Id of the file
         type: integer
         required: True
+      - name: usertoken
+        in: header
+        description: authentication token of user
+        required: true
+        type: string
     responses:
       200:
         description: Successfully found and delievered file
@@ -230,22 +270,31 @@ def get_single_file(id):
       404:
           description: Not Found
     """
-    file = File1.query.get(id)
+    username = get_jwt_identity()
+    if not username:
+        return not_authenticated()
+    file = File1.query.get(file_id)
     if file:
+        if file.user != username:
+            return not_authorized()
         js = json.dumps(file, cls=FileEncoder)
         return Response(js, status=200, mimetype='application/json')
     else:
         return not_found()
 
 
-@app.route("/files/<id>/download", methods=["GET"])
-def get_download_link(id):
-    # StoredFile.public_url
+@app.route("/files/<file_id>/download", methods=["GET"])
+@jwt_required
+def get_download_link(file_id):
+    username = get_jwt_identity()
+    if not username:
+        return not_authenticated()
+    # todo StoredFile.public_url
     message = {
         "status": 404,
         "message": "Not yet implemented",
     }
-    return Response(json.dumps(message), status=404)
+    return Response(json.dumps(message), status=404, mimetype='application/json')
 
 
 @app.errorhandler(404)
@@ -254,7 +303,7 @@ def not_found(error=None):
         "status": 404,
         "message": "Not Found: " + request.url,
     }
-    return Response(json.dumps(message), status=404)
+    return Response(json.dumps(message), status=404, mimetype='application/json')
 
 
 @app.errorhandler(401)
@@ -263,15 +312,19 @@ def not_authenticated(error=None, msg="Failed authentication!"):
         "status": 401,
         "message": "Unauthorized: " + msg,
     }
-    return Response(json.dumps(message), status=401)
+    return Response(json.dumps(message), status=401, mimetype='application/json')
 
 
-def authenticate(usertoken):
-    # todo add functionality
-    return True
+@app.errorhandler(403)
+def not_authorized(error=None):
+    message = {
+        "status": 403,
+        "message": "Forbidden: Insufficient access rights",
+    }
+    return Response(json.dumps(message), status=403, mimetype='application/json')
 
 
-# TODO add test data
+# add test data
 # db.session.add(File1("filename", 100, "user"))
 # db.session.commit()
 # db.session.add(File1("filename2", 1000, "user"))
